Create a store hook to hide redux stuff
